% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bk_lookup_chin}
\alias{bk_lookup_chin}
\title{Lookup table for Chinook stock IDs from BkFRAM to normal FRAM}
\format{
A data frame with 116 rows and 3 columns:
\describe{
  \item{stock_name_bk}{Stock name}
  \item{bk_stock_id }{Stock ID in the BackwardsFRAM table}
  \item{stock_id}{Stock ID anywhere else in FRAM and TAMM}
}
}
\usage{
bk_lookup_chin
}
\description{
This table is based on Chinook stock tables from pre-season 2024. Additional stock may
be added in the future (splitting out stock aggregates). The `bk_lookupfun_chin()` function accommodates
this.
}
\keyword{datasets}
