% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bk_lookupfun.R
\name{bk_lookupfun_chin}
\alias{bk_lookupfun_chin}
\title{Generate Chinook bk_fram look-up table}
\usage{
bk_lookupfun_chin(max_stock = NA)
}
\arguments{
\item{max_stock}{Positive integer representing the largest *Normal* stock id (stock ids outside of the BackwardsFram table). As of Preseason 2024, there are 78 stocks; splitting stocks might increase this (e.g. 80, 82, 84). Defaults to NA.}
}
\value{
Tibble of lookup table. `$stock_name_bk` gives numan-readable stock names; `$bk_stock_id` gives the "StockID" used
in backwards FRAM; `$stock_id` gives the StockID used everywhere else. NAs in `$stock_id` correspond to "totalled" stocks, which are used in
backwards FRAM and nowhere else.
}
\description{
At present the FRAM databases use a differents stock_id numbering system in the BackwardsFRAM
 table than everywhere else. The dataframe `bk_lookup_chin` provides a lookup table based on the stocks
 used in Preseason 2024. However, we are working to modify FRAM to support splitting Chinook stock aggregates, which
 will lead to increased numbers of StockIDs. `bk_lookupfun_chin` provides an appropruate lookup table based on the `max_stock` argument, either
 returning `bk_lookup_chin` (if max_stock is not provided or doesn't exceed the stocks in `bk_lookup_chin`) or extending the dataframe
 with new stock ids based on the numbering conventions of the bk_stock_id.
}
\examples{
temp = bk_lookupfun_chin(80)
tail(temp)
}
