% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrity.R
\name{clean_species}
\alias{clean_species}
\title{Clean variations on function names}
\usage{
clean_species(species, call = rlang::caller_env())
}
\arguments{
\item{species}{character atomic}

\item{call}{argument to ensure error messages correctly point to the calling function rather than this one.}
}
\value{
either "COHO" or "CHINOOK"
}
\description{
Takes a character string with putative species name corresponding to "CHINOOK" or "COHO" and returns exactly "CHINOOK" or "COHO".
Works by matching against predefined variations (different capitalizations, "chin" instead of "chinook"). Returns informative error if
species name does not match any of these variations. Intended for internal use, to streamline allowing variations of species in arguments.
}
