% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fishery_coho_psc}
\alias{fishery_coho_psc}
\title{Coho fishery information (PSC)}
\format{
A data frame with 197 rows and 5 columns:
\describe{
  \item{species}{Species name}
  \item{fram_fishery_id}{Fishery ID in FRAM}
  \item{psc_fishery_id}{Corresponding ID for the psc fishery}
  \item{psc_group_code}{}
  \item{psc_fishery_name}{Name of PSC fishery ID for the psc fishery}
  \item{psc_fishery_order}{}
}
}
\source{
https://github.com/PSC-CoTC/AnnualReport
}
\usage{
fishery_coho_psc
}
\description{
Mapping of FRAM fishery_id to PSC fishery ID and name for Coho salmon, based on the AnnualReport R package.
}
\keyword{datasets}
