% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{fishery_search}
\alias{fishery_search}
\alias{stock_search}
\title{Search through fishery and stock lookup tables.}
\usage{
fishery_search(pattern, species)

stock_search(pattern, species)
}
\arguments{
\item{pattern}{Character string of pattern to search for, case insensitive. If numeric instead, function will return the row with the corresponding stock or fishery ID.}

\item{species}{"CHINOOK" or "COHO" (or lowercase permutations)}
}
\description{
Search through fishery and stock lookup tables.
}
\examples{
fishery_search("kmz", "COHO")
fishery_search(50, "COHO")
stock_search("nooksack", "CHINOOK")
stock_search(21, "CHINOOK")
}
