% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_columns.R
\name{label_fisheries}
\alias{label_fisheries}
\title{Add `fishery_label` column to describe fisheries of dataframe}
\usage{
label_fisheries(.data, species = NULL, fisheries_col = "fishery_id")
}
\arguments{
\item{.data}{Dataframe containing `fishery_id` column (or analogous column with different name specied by `fisheries_col` argument)}

\item{species}{character atomic identifying either "CHINOOK" or "COHO" (or variations of those). Defaults to `NULL`; only needed if `.data` does not have species defined as an attribute. Tables read from a fram database using `framrsquared` package automatically have appropriate species attribute.}

\item{fisheries_col}{character atomic identifying the column treated as containing fishery_id values Defaults to `fishery_id`.}
}
\value{
Data frame with additional column `fishery_label`
}
\description{
Streamlined version of using `left_join`, `rename`, and `relocate` to add human-readable text to a dataframe with fishery_id or similar columns.
Automatically detects species and `fishery_id` column; user can specify species or alternative name for fishery_id column if needed.
}
\examples{
## create example data
dat = data.frame(fishery_id = sample(1:73, size = 10, replace = TRUE))
## Add fishery label
dat |> label_fisheries(species = "CHINOOK")

}
