% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_columns.R
\name{label_stocks}
\alias{label_stocks}
\title{Add `stock_label` column to describe fisheries of dataframe}
\usage{
label_stocks(.data, species = NULL, stocks_col = "stock_id")
}
\arguments{
\item{.data}{Dataframe containing `stock_id` column (or analogous column with different name specied by `stocks_col` argument)}

\item{species}{character atomic identifying either "CHINOOK" or "COHO" (or variations of those). Defaults to `NULL`; only needed if `.data` does not have species defined as an attribute. Tables read from a fram database using `framrsquared` package automatically have appropriate species attribute.}

\item{stocks_col}{character atomic identifying the column treated as containing stock_id values Defaults to `stock_id`.}
}
\value{
Data frame with additional column `stock_label`
}
\description{
Streamlined version of using `left_join`, `rename`, and `relocate` to add human-readable text to a dataframe with stock_id or similar columns.
Automatically detects species and `stock_id` column; user can specify species or alternative name for stock_d column if needed.
}
\examples{
## create example data
dat = data.frame(stock_id = sample(1:73, size = 10, replace = TRUE))
## Add stock information
dat |> label_stocks(species = "CHINOOK")

}
