% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{limiting_stock_coho}
\alias{limiting_stock_coho}
\title{Coho limiting stock}
\format{
A tibble with 10 rows and 2 columns:
\describe{
  \item{name}{Name of the limiting stock}
  \item{name_tamm}{Exact match of limiting stock name in the TAMM}
  \item{stock_id_vec}{a list of numeric vectors, each containing all fram stock_id values for that TAMM limiting stock}
}
#' @source 2024 Pre-Season Chinook DB.mdb
}
\usage{
limiting_stock_coho
}
\description{
Maps between the limiting stock of the Coho TAMM (e.g., 'Hood Canal Wild') and FRAM stock ids. Useful for summarizing impacts on stocks of concern.
}
\keyword{datasets}
