% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rmis_fram}
\alias{rmis_fram}
\title{Lookup table for FRAM to RMIS recoveries}
\format{
## `rmis_fram`
A data frame with 225,390 rows and 4 columns:
\describe{
  \item{psc_code}{RMIS PSC Recovery Location}
  \item{gear}{Gear Code}
  \item{fishery_id}{FRAM fishery ID}
  \item{fishery_name}{FRAM fishery name}
}
}
\source{
<https://github.com/FRAMverse/framr/blob/master/xlsx/lu_coho.xlsx
}
\usage{
rmis_fram
}
\description{
For more about the Regional Mark Information System (RMIS), see https://www.rmpc.org/.
}
\keyword{datasets}
