% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stock_coho_fram}
\alias{stock_coho_fram}
\title{Coho stock information}
\format{
A data frame with 78 rows and 7 columns:
\describe{
  \item{species}{Species name}
  \item{stock_version}{}
  \item{stock_id}{Coho stock id number in FRAM}
  \item{production_region_number}{}
  \item{management_unit_number}{}
  \item{stock_name}{Coho stock name in FRAM}
  \item{stock_long_name}{`stock_name` but more human readable}
  \item{limiting_stock_name}{Identifier for which of the limiting stock names of the Coho TAMM (sheet `2`) this stock contributes to, if any.}
  #'   \item{limiting_stock_name_tamm}{As `limiting_stock_name`, but exact match to name in TAMM.}
}
}
\source{
2024NOF_CohoFRAMdatabase_distribution.mdb
}
\usage{
stock_coho_fram
}
\description{
Mapping of FRAM stock_id to FRAM stock name for Coho salmon, taken from the `Stock` table of
the Coho FRAM database.
}
\keyword{datasets}
