% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stock_coho_psc}
\alias{stock_coho_psc}
\title{Coho stock information (PSC)}
\format{
A data frame with 41 rows and 4 columns:
\describe{
  \item{species}{Species name}
  \item{fram_stock_id}{Coho stock id number in FRAM}
  \item{psc_stock_id}{Corresponding PSC stock ID}
  \item{psc_stock_name}{PSC stock name}
}
}
\source{
https://github.com/PSC-CoTC/AnnualReport
}
\usage{
stock_coho_psc
}
\description{
Mapping of FRAM stock_id to PSC stock id and psc stock names, based on the AnnualReport R package.
}
\keyword{datasets}
