% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tocas_fram}
\alias{tocas_fram}
\title{Lookup table for FRAM to TOCAS}
\format{
## `tocas_fram`
A data frame with 10,462 rows and 6 columns:
\describe{
  \item{fisher_type}{1 = Non-Treaty / 2 = Treaty}
  \item{gear}{Gear Code}
  \item{dispostion}{Dispostion. "COMM" = Commercial, "C&SF" = Ceremonial and Subsistence fishing, "ORGN" = , "Test" and "TEST" = Test fishery (?), TKHM = ??}
  \item{catch_area}{Catch area code, not zero-padded}
  \item{fishery_id}{FRAM fishery ID}
  \item{fishery_name}{FRAM fishery name}
}
}
\source{
<https://github.com/FRAMverse/framr/blob/master/xlsx/lu_coho.xlsx
}
\usage{
tocas_fram
}
\description{
Lookup table for FRAM to TOCAS
}
\keyword{datasets}
