% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{waft_fram}
\alias{waft_fram}
\title{Lookup table for FRAM to WAFT}
\format{
## `waft_fram`
A data frame with 10,462 rows and 6 columns:
\describe{
  \item{fishery_type}{1 = Non-Treaty / 2 = Treaty}
  \item{gear} {Gear Code}
  \item{dispostion}{Dispostion - Commerical, Ceremonial Etc...}
  \item{catch_area}{Catch area code, zero-padded}
  \item{fishery_id}{FRAM fishery ID}
  \item{fishery_name} {FRAM fishery name}
}
}
\source{
<https://github.com/FRAMverse/framr/blob/master/xlsx/lu_coho.xlsx
}
\usage{
waft_fram
}
\description{
Lookup table for FRAM to WAFT
}
\keyword{datasets}
